synergies -> data
